+++ CHANGELOG.md
@@ -5,8 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.13.0] - 2026-02-06
 ### Added
 - Added "On video tap" option to choose between in-app player or default player app ([#917])
+
+### Changed
+- Updated translations
 
 ### Fixed
 - Fixed unnecessary "Video has no audio" toast when looping videos ([#876])
@@ -299,7 +304,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#917]: https://github.com/FossifyOrg/Gallery/issues/917
 [#925]: https://github.com/FossifyOrg/Gallery/issues/925
 
-[Unreleased]: https://github.com/FossifyOrg/Gallery/compare/1.12.0...HEAD
+[Unreleased]: https://github.com/FossifyOrg/Gallery/compare/1.13.0...HEAD
+[1.13.0]: https://github.com/FossifyOrg/Gallery/compare/1.12.0...1.13.0
 [1.12.0]: https://github.com/FossifyOrg/Gallery/compare/1.11.0...1.12.0
 [1.11.0]: https://github.com/FossifyOrg/Gallery/compare/1.10.0...1.11.0
 [1.10.0]: https://github.com/FossifyOrg/Gallery/compare/1.9.1...1.10.0